requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
google-generativeai
langchain-chroma
pypdf
chromadb
pydantic
python-multipart
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
